spin_app
Struct App

pub struct App { /* private fields */ }

An App holds loaded configuration for a Spin application.

Implementations

impl App

pub fn new(id: impl Into<String>, locked: LockedApp) -> Self

Returns a new app for the given runtime-specific identifier and locked app.

pub fn id(&self) -> &str

Returns a runtime-specific identifier for this app.

pub fn get_metadata<'this, T: Deserialize<'this>>( &'this self, key: MetadataKey<T>, ) -> Result<Option<T>>

Deserializes typed metadata for this app.

Returns Ok(None) if there is no metadata for the given key and an Err only if there is a value for the key but the typed deserialization failed.

pub fn require_metadata<'this, T: Deserialize<'this>>( &'this self, key: MetadataKey<T>, ) -> Result<T>

Deserializes typed metadata for this app.

Like App::get_metadata, but returns an error if there is no metadata for the given key.

pub fn variables(&self) -> impl Iterator<Item = (&String, &Variable)>

Returns an iterator of custom config Variables defined for this app.

pub fn components(&self) -> impl Iterator<Item = AppComponent<'_>>

Returns an iterator of AppComponents defined for this app.

pub fn get_component(&self, component_id: &str) -> Option<AppComponent<'_>>

Returns the AppComponent with the given component_id, or None if it doesnโt exist.

pub fn triggers(&self) -> impl Iterator<Item = AppTrigger<'_>> + '_

Returns an iterator of AppTriggers defined for this app.

pub fn get_trigger_metadata<'this, T: Deserialize<'this>>( &'this self, trigger_type: &str, ) -> Result<Option<T>>

Returns the trigger metadata for a specific trigger type.

pub fn triggers_with_type<'a>( &'a self, trigger_type: &'a str, ) -> impl Iterator<Item = AppTrigger<'_>>

Returns an iterator of AppTriggers defined for this app with the given trigger_type.

pub fn trigger_configs<'a, T: Deserialize<'a>>( &'a self, trigger_type: &'a str, ) -> Result<impl IntoIterator<Item = (&'a str, T)>>

Returns an iterator of trigger IDs and deserialized trigger configs for the given trigger_type.

pub fn ensure_needs_only(&self, supported: &[&str]) -> Result<(), String>

Checks that the application does not have any host requirements outside the supported set. The error case returns a comma-separated list of unmet requirements.