Crate spin_app

Spin internal application interfaces

This crate contains interfaces to Spin application configuration to be used by crates that implement Spin execution environments: trigger executors and host components, in particular.

Modules

locked

Spin lock file (spin.lock) serialization models.

values

Dynamically-typed value helpers.

Structs

App

An App holds loaded configuration for a Spin application.

AppComponent

An AppComponent holds configuration for a Spin application component.

AppTrigger

An AppTrigger holds configuration for a Spin application trigger.

MetadataKey

MetadataKey is a handle to a typed metadata value.

Variable

A Variable specifies a custom configuration variable.

Enums

Error

Errors returned by methods in this crate.

Constants

APP_DESCRIPTION_KEY

MetadataKey for extracting the application description.

APP_NAME_KEY

MetadataKey for extracting the application name.

APP_VERSION_KEY

MetadataKey for extracting the application version.

OCI_IMAGE_DIGEST_KEY

MetadataKey for extracting the OCI image digest.

Functions

retain_components

Scrubs the locked app to only contain the given list of components Introspects the LockedApp to find and selectively retain the triggers that correspond to those components

Type Aliases

Result

Type alias for a Results with Error.

ValidatorFn

Validation function type for ensuring that applications meet requirements even with components filtered out.