Crate spin_app

Spin internal application interfaces

This crate contains interfaces to Spin application configuration to be used by crates that implement Spin execution environments: trigger executors and host components, in particular.

Modules

locked

Spin lock file (spin.lock) serialization models.

values

Dynamically-typed value helpers.

Structs

App

An App holds loaded configuration for a Spin application.

AppComponent

An AppComponent holds configuration for a Spin application component.

AppLoader

An AppLoader holds an implementation of Loader along with DynamicHostComponents configuration.

AppTrigger

An AppTrigger holds configuration for a Spin application trigger.

MetadataKey

MetadataKey is a handle to a typed metadata value.

OwnedApp

The self-referencing struct.

Variable

A Variable specifies a custom configuration variable.

Enums

Error

Errors returned by methods in this crate.

Constants

APP_DESCRIPTION_KEY

MetadataKey for extracting the application description.

APP_NAME_KEY

MetadataKey for extracting the application name.

APP_VERSION_KEY

MetadataKey for extracting the application version.

OCI_IMAGE_DIGEST_KEY

MetadataKey for extracting the OCI image digest.

Traits

DynamicHostComponent

A trait for “dynamic” Spin host components.

Loader

A trait for implementing the low-level operations needed to load an App.

Type Aliases

Result

Type alias for a Results with Error.

Attribute Macros

async_trait