Crate spin_app

Spin internal application interfaces

This crate contains interfaces to Spin application configuration to be used by crates that implement Spin execution environments: trigger executors and host components, in particular.

Re-exports

pub use locked::Variable;

Modules

locked

Spin lock file (spin.lock) serialization models.

values

Dynamically-typed value helpers.

Structs

App

An App holds loaded configuration for a Spin application.

AppComponent

An AppComponent holds configuration for a Spin application component.

AppLoader

An AppLoader holds an implementation of Loader along with DynamicHostComponents configuration.

AppTrigger

An AppTrigger holds configuration for a Spin application trigger.

OwnedApp

The self-referencing struct.

Enums

Error

Errors returned by methods in this crate.

Traits

DynamicHostComponent

A trait for “dynamic” Spin host components.

Loader

A trait for implementing the low-level operations needed to load an App.

Type Definitions

Result

Type alias for a Results with Error.

Attribute Macros

async_trait